Don't do partial matches (e.g. pt-PT -> pt-BR)
If we have a language that is good for all locales, we can name that
language by just language (e.g. 'pt'). What we should NOT do is match
'pt-PT' to 'pt-BR' in the absence of 'pt'.

This is apparently desired behavior in the current jed/locale, so we'll
fork.